restructure tests and sort out failing test.
