+++ src/SFA.DAS.PSRService.Web.UnitTests/ServiceTests/ReportServiceTests/Can_Be_Edited/Given_A_Not_Submitted_Report/And_Report_Is_For_Current_Period/And_Report_Is_For_Current_Period.cs
@@ -2,13 +2,13 @@
 using Moq;
 using SFA.DAS.PSRService.Domain.Entities;
 
-namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Given_A_Not_Submitted_Report.And_Report_Is_Not_For_Current_Period
+namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Can_Be_Edited.Given_A_Not_Submitted_Report.And_Report_Is_For_Current_Period
 {
     [ExcludeFromCodeCoverage]
-    public abstract class And_Report_Is_Not_For_Current_Period
-        :Given_A_Not_Submitted_Report
+    public abstract class And_Report_Is_For_Current_Period
+    :Given_A_Not_Submitted_Report
     {
-        public And_Report_Is_Not_For_Current_Period()
+        public And_Report_Is_For_Current_Period()
         {
             MockPeriodService
                 .Setup(
@@ -16,7 +16,7 @@ namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Given
                         m.ReportIsForCurrentPeriod(
                             It.IsAny<Report>()
                         ))
-                .Returns(false);
+                .Returns(true);
         }
     }
 }
