+++ src/SFA.DAS.PSRService.Web.UnitTests/ServiceTests/ReportServiceTests/Can_Be_Edited/Given_A_Not_Submitted_Report/And_Report_Is_For_Current_Period/When_ReportService_CanBeEdited_Is_Called.cs
@@ -1,16 +1,14 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using FluentAssertions;
-using Moq;
 using NUnit.Framework;
-using SFA.DAS.PSRService.Domain.Entities;
 
-namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.
-    Given_A_Null_Report
+namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Can_Be_Edited.Given_A_Not_Submitted_Report.
+    And_Report_Is_For_Current_Period
 {
     [ExcludeFromCodeCoverage]
     [TestFixture]
     public class When_ReportService_CanBeEdited_Is_Called
-        : Given_A_Null_Report
+        : And_Report_Is_For_Current_Period
     {
         private bool response;
 
@@ -24,19 +22,11 @@ namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.
         }
 
         [Test]
-        public void Then_False_Is_Returned()
+        public void Then_True_Is_Returned()
         {
             response
                 .Should()
-                .BeFalse();
-        }
-
-        [Test]
-        public void Then_PeriodService_ReportIsForCurrentPeriod_Is_Not_Called()
-        {
-            MockPeriodService
-                .Verify( 
-                    m => m.ReportIsForCurrentPeriod(It.IsAny<Report>()), Times.Never);
+                .BeTrue();
         }
     }
 }
