+++ src/SFA.DAS.PSRService.Web.UnitTests/ServiceTests/ReportServiceTests/Can_Be_Edited/Given_A_Not_Submitted_Report/And_Report_Is_Not_For_Current_Period/When_ReportService_CanBeEdited_Is_Called.cs
@@ -2,13 +2,12 @@
 using FluentAssertions;
 using NUnit.Framework;
 
-namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Given_A_Submitted_Report.
-    And_Report_Is_For_Current_Period
+namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Can_Be_Edited.Given_A_Not_Submitted_Report.And_Report_Is_Not_For_Current_Period
 {
     [ExcludeFromCodeCoverage]
     [TestFixture]
     public class When_ReportService_CanBeEdited_Is_Called
-        : And_Report_Is_For_Current_Period
+    : And_Report_Is_Not_For_Current_Period
     {
         private bool response;
 
@@ -16,9 +15,9 @@ namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Given
         {
             response
                 =
-                SUT
-                    .CanBeEdited(
-                        StubReport);
+            SUT
+                .CanBeEdited(
+                    StubReport);
         }
 
         [Test]
