+++ src/SFA.DAS.PSRService.Web.UnitTests/ServiceTests/ReportServiceTests/Can_Be_Edited/Given_A_Not_Submitted_Report/Given_A_Not_Submitted_Report.cs
@@ -5,16 +5,16 @@ using SFA.DAS.PSRService.Domain.Entities;
 using SFA.DAS.PSRService.Web.Configuration;
 using SFA.DAS.PSRService.Web.Services;
 
-namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Given_A_Null_Report
+namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Can_Be_Edited.Given_A_Not_Submitted_Report
 {
     [ExcludeFromCodeCoverage]
-    public abstract class Given_A_Null_Report
-    :GivenWhenThen<IReportService>
+    public abstract class Given_A_Not_Submitted_Report
+        : GivenWhenThen<IReportService>
     {
         protected Mock<IPeriodService> MockPeriodService;
         protected Report StubReport;
 
-        public Given_A_Null_Report()
+        public Given_A_Not_Submitted_Report()
         {
             MockPeriodService = new Mock<IPeriodService>();
 
@@ -24,7 +24,7 @@ namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Given
                 periodService: MockPeriodService.Object
             );
 
-            StubReport = null;
+            StubReport = new Report {Submitted = false};
         }
     }
 }
