+++ src/SFA.DAS.PSRService.Web.UnitTests/ServiceTests/ReportServiceTests/Can_Be_Edited/Given_A_Submitted_Report/Given_A_Submitted_Report.cs
@@ -5,7 +5,7 @@ using SFA.DAS.PSRService.Domain.Entities;
 using SFA.DAS.PSRService.Web.Configuration;
 using SFA.DAS.PSRService.Web.Services;
 
-namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Given_A_Submitted_Report
+namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Can_Be_Edited.Given_A_Submitted_Report
 {
     [ExcludeFromCodeCoverage]
     public class Given_A_Submitted_Report
