+++ src/SFA.DAS.PSRService.Web.UnitTests/ServiceTests/ReportServiceTests/Submit/Given_A_Submitted_Report/Given_A_Submitted_Report.cs
@@ -1,0 +1,26 @@
+﻿using System.Diagnostics.CodeAnalysis;
+using MediatR;
+using Moq;
+using SFA.DAS.PSRService.Domain.Entities;
+using SFA.DAS.PSRService.Web.Configuration;
+using SFA.DAS.PSRService.Web.Services;
+
+namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Submit.Given_A_Submitted_Report
+{
+    [ExcludeFromCodeCoverage]
+    public abstract class Given_A_Submitted_Report
+    :GivenWhenThen<IReportService>
+    {
+        protected Report AlreadySubmittedReport;
+
+        public Given_A_Submitted_Report()
+        {
+            SUT = new ReportService(
+                config: Mock.Of<IWebConfiguration>(),
+                mediator: Mock.Of<IMediator>(),
+                periodService: Mock.Of<IPeriodService>());
+
+            AlreadySubmittedReport = new ReportBuilder().BuildAlreadySubmittedReportWithValidSections();
+        }      
+    }
+}
