+++ src/SFA.DAS.PSRService.Web.UnitTests/ServiceTests/ReportServiceTests/Submit/Given_A_Valid_Report/And_Report_Is_Not_Submitted.cs
@@ -1,0 +1,14 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace SFA.DAS.PSRService.Web.UnitTests.ServiceTests.ReportServiceTests.Submit.Given_A_Valid_Report
+{
+    [ExcludeFromCodeCoverage]
+    public abstract class And_Report_Is_Not_Submitted
+    : Given_A_Valid_Report
+    {
+        public And_Report_Is_Not_Submitted()
+        {
+            ValidNotSubmittedReport.Submitted = false;
+        }
+    }
+}
